Generalize drop and create options
